fix: harden useDisclosure state tracking
