+++ tests/use-disclosure.test.tsx
@@ -1,5 +1,4 @@
 import { act, renderHook } from "@testing-library/react";
-import { renderToString } from "react-dom/server";
 import { describe, expect, it, vi } from "vitest";
 
 import { useDisclosure } from "../src/client/hooks";
@@ -233,17 +232,5 @@ describe("useDisclosure", () => {
     expect(result.current.close).toBe(firstClose);
     expect(result.current.toggle).toBe(firstToggle);
     expect(result.current.setOpen).toBe(firstSetOpen);
-  });
-
-  it("renders safely on the server", () => {
-    function ServerRenderedDisclosure() {
-      const { isOpen } = useDisclosure({
-        defaultOpen: true,
-      });
-
-      return <span>{String(isOpen)}</span>;
-    }
-
-    expect(renderToString(<ServerRenderedDisclosure />)).toContain("<span>true</span>");
   });
 });
